requirements.txt
flask==3.0.3
groq==0.9.0
chromadb==0.5.3
sentence-transformers==3.0.1
flask-limiter==3.7.0
redis==5.0.8
requests==2.32.3
python-dotenv==1.0.1